+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJAQ4HT7KRSYVAC65G05TX3M"
+        "01KJATQJ2ZQ02474QA3MKRJ4A9"
     )
     r"""The unique identifier of the data source"""