+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJAQ4HTBEFBMAHKY9BM2K4WE"
+        "01KJATQJ35ZHMZEN75MPVBQNJX"
     )
     r"""The unique identifier of the data source"""