+++ packages/orq-rc/RELEASES.md
@@ -6129,3 +6129,13 @@ Based on:
 - [python v4.5.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.7 - packages/orq-rc
+
+## 2026-02-25 16:42:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.729.0 (2.841.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.8 - packages/orq-rc